requirements.txt
torch
numpy
transformers
datasets
tiktoken
tqdm
pyyaml
